refactor: extract EmailCTA to client component
