+++ public/sitemap.xml
@@ -4,7 +4,7 @@
   
   <url>
     <loc>https://gauresh.is-a.dev/</loc>
-    <lastmod>2026-03-22T13:54:54.691Z</lastmod>
+    <lastmod>2026-03-22T14:12:08.625Z</lastmod>
     <changefreq>weekly</changefreq>
     <priority>1</priority>
     
@@ -46,14 +46,14 @@
   </url>
   <url>
     <loc>https://gauresh.is-a.dev/about/</loc>
-    <lastmod>2026-03-22T13:54:54.691Z</lastmod>
+    <lastmod>2026-03-22T14:12:08.625Z</lastmod>
     <changefreq>monthly</changefreq>
     <priority>0.8</priority>
     
   </url>
   <url>
     <loc>https://gauresh.is-a.dev/projects/</loc>
-    <lastmod>2026-03-22T13:54:54.691Z</lastmod>
+    <lastmod>2026-03-22T14:12:08.625Z</lastmod>
     <changefreq>weekly</changefreq>
     <priority>0.9</priority>
     
@@ -200,7 +200,7 @@
   </url>
   <url>
     <loc>https://gauresh.is-a.dev/works/</loc>
-    <lastmod>2026-03-22T13:54:54.691Z</lastmod>
+    <lastmod>2026-03-22T14:12:08.625Z</lastmod>
     <changefreq>weekly</changefreq>
     <priority>0.9</priority>
     
@@ -282,7 +282,7 @@
   </url>
   <url>
     <loc>https://gauresh.is-a.dev/certificates/</loc>
-    <lastmod>2026-03-22T13:54:54.691Z</lastmod>
+    <lastmod>2026-03-22T14:12:08.625Z</lastmod>
     <changefreq>monthly</changefreq>
     <priority>0.7</priority>
